+++ src/js-framework/lib/__test__/test.js
@@ -369,6 +369,22 @@ describe('test input and output', function () {
     delete allDocs[name]
   })
 
+  it('computed in repeat case', function () {
+    var name = 'computed-in-repeat'
+    var inputCode = readInput(name)
+    var outputCode = readOutput(name)
+    var doc = new Document(name)
+    allDocs[name] = doc
+
+    framework.createInstance(name, inputCode)
+    var expected = eval('(' + outputCode + ')')
+    var actual = doc.toJSON()
+    expect(actual).eql(expected)
+
+    framework.destroyInstance(name)
+    delete allDocs[name]
+  })
+
   it('backward(register/render) case', function () {
     var name = 'backward1'
     var inputCode = readInput(name)