+++ src/js-framework/CHANGELOG.md
@@ -4,3 +4,4 @@
 
 - add `updateFinish` action after `click/fireEvent`
 - add a testcase for `track-by`
+- fixed #323. fix computed in repeat scope